requirements.txt
Cython==3.1.4
einops==0.8.1
lightning==2.5.5
numpy==2.3.3
prodigyopt==1.1.2
pygame==2.6.1
PyOpenGL==3.1.10
PyYAML==6.0.2
soundfile==0.13.1
torch==2.8.0
torchaudio==2.8.0
